==> R/00_scratch.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r}
library(tidyverse)
library(ggmap)
dat <- read_rds("dat.rds")

test <- dat %>% 
  filter(!is.na(WellLocation), !is.na(City)) %>% 
  dplyr::select(WellLocation, City) %>% 
  mutate(address = paste(WellLocation, City, "CA", sep = ", "))

library(ggmap)
geocode(test[1:100,]$address, output = "latlona", source = "google")
gc <- .Last.value

gc2 <- geocode(test[200000:200020,]$address, output = "latlona", source = "google")

gc3 <- geocode(test[300000:300020,]$address, output = "latlona", source = "google")
```

Internals for geocoding user input addresses and linking them to the map. Should be triggered by a button click.
```{r}
input_street <- "647 D St"
input_city <- "Davis"
input_address <- paste(input_street, input_city, sep = ", ")

output_ll <- ggmap::geocode(input_address, output = "latlon", source = "google")
output_ll
```

Leaflet showing GW levels. 
```{r}
library(raster)
library(leaflet)

# list of interpolations
ml <- read_rds("ml.rds")

# ensemble predicitons
ensemble <- lapply(1:12, function(x){ ml[[x]]$Ensemble }) 

# projected for leaflet
ep <- lapply(ensemble, leaflet::projectRasterForLeaflet, method = "bilinear") 

# raster with max value sets the palette
maxr <- sapply(ensemble, function(x){x@data@values %>% max(na.rm = T)}) %>% which.max()

# make palette
pal <- colorNumeric(colormap(colormaps$jet, nshades = 10), values(ensemble[[maxr]]),
                    na.color = "transparent")

# make titles 
g <- paste0(rep(c("SP","FA"), 7), rep(11:17, each =2))[-c(2,4)]


# make leaflet
leaflet() %>% addProviderTiles(provider = providers$CartoDB) %>% 
  addRasterImage(ep[[1]], colors = pal, opacity = 0.7, project = FALSE, group = g[1]) %>%
  addRasterImage(ep[[2]], colors = pal, opacity = 0.7, project = FALSE, group = g[2]) %>%
  addRasterImage(ep[[3]], colors = pal, opacity = 0.7, project = FALSE, group = g[3]) %>%
  addRasterImage(ep[[4]], colors = pal, opacity = 0.7, project = FALSE, group = g[4]) %>%
  addRasterImage(ep[[5]], colors = pal, opacity = 0.7, project = FALSE, group = g[5]) %>%
  addRasterImage(ep[[6]], colors = pal, opacity = 0.7, project = FALSE, group = g[6]) %>%
  addRasterImage(ep[[7]], colors = pal, opacity = 0.7, project = FALSE, group = g[7]) %>%
  addRasterImage(ep[[8]], colors = pal, opacity = 0.7, project = FALSE, group = g[8]) %>%
  addRasterImage(ep[[9]], colors = pal, opacity = 0.7, project = FALSE, group = g[9]) %>%
  addRasterImage(ep[[10]], colors = pal, opacity = 0.7, project = FALSE, group = g[10]) %>%
  addRasterImage(ep[[11]], colors = pal, opacity = 0.7, project = FALSE, group = g[11]) %>%
  addRasterImage(ep[[12]], colors = pal, opacity = 0.7, project = FALSE, group = g[12]) %>%
  addLayersControl(baseGroups = g, position = "topleft") %>% 
  # raster with max val sets legend scale
  addLegend(pal = pal, values = values(ensemble[[maxr]]), 
            title = "Depth to GW (ft)") %>% 
  addEasyButton(easyButton(
    icon="fa-globe", title="Zoom to Level 6",
    onClick=JS("function(btn, map){ map.setZoom(6); }"))) %>%
  addEasyButton(easyButton(
    icon="fa-crosshairs", title="Locate Me",
    onClick=JS("function(btn, map){ map.locate({setView: true}); }"))) %>% 
  addMarkers(lat = output_ll$lat, lng = output_ll$lon)
```


Calculate confidence intervals around kirging estimate with kriging variance.
```{r}
# get kriging variances. Seem WAY too low
var <- lapply(1:12, function(x){ml[[x]]$OK_var})

# plot OK estimate and 95% CIs
par(mfrow = c(1,3))
  # dividing by sample size (n = 2246) but only makes variance smaller... not doing much....
  (ml[[1]]$OK - (sqrt(var[[1]]/2246) * 1.96)) %>% plot(main = "2.5%")
  ml[[1]]$OK  %>% plot(main = "mean")
  (ml[[1]]$OK + (sqrt(var[[1]]/2246) * 1.96)) %>% plot(main = "97.5%")
```
