==> R/10_space_time_kriging.Rmd <==
---
title: "Space Time Kirging"
output: html_document
---

The purpose of this script is to implement space time kirging on the seasonal GW levels obtained from DWR's unconfined to semiconfined netowrk of monitoring wells.

Packages used.
```{r}
library(gstat)
library(sp)
library(raster)
library(dplyr)
library(readr)
library(ggplot2)
library(lubridate)
library(here)      # load here after lubridate to avoid `here` conflict
library(spacetime)
library(lattice)
```

Read in all shapefiles, and CV basin boundary.
```{r}
# list of seasonal, log-transformed spatial points in the CV
dl_cv_ln <- read_rds("dl_cv_log_trans.rds")

# cental valley alluval basin boundary
cv <- shapefile(here("data","spatial",
                     "central_valley_alluvial_boundary",
                     "Alluvial_Bnd.shp"))
cv <- spTransform(cv, crs(dl_cv_ln[[1]]))
```

We further simplify the problem by assuming that all points within a "season," which may refelct a 3-4 month time period, come from 1 time step, denoted by the season. For example, all points sampled within Spring 2011 are assumed to have the same time step. This simplifies the number of time frames in the study to the number of seasons.  
```{r}
# # make time slots
# t <- ymd_hms(paste(paste(rep(2011:2017, each=2), c("02","08"), "01", sep = "-"), 
#                    "00:00:00")[-c(2,4)])
# 
# # replicate time slots to match number of observations per time slot
# 
# ## number of observations per time period
# n <- sapply(dl_cv_ln, function(x){return(nrow(x))})
# 
# ## initalize vector of replicated time slots
# tr <- vector("list", length = length(t))  
# 
# ## loop to create replicated time slots
# for(i in 1:length(t)){
#   tr[[i]] <- rep(t[i], times = n[i])
# }
# 
# ## bind into vector
# tr <- do.call(c, tr)


# use explicit dates
tt <- lapply(dl_cv_ln, 
             function(x){return(mutate(x@data, 
                                       md = ymd_hms(paste(Msmt_Date, "00:00:00"))) %>% 
                                pull(md))}) %>% 
      do.call(c, .)

# extract spatial points into one spatial points object with defined crs
sp_pts <- SpatialPoints(do.call(rbind, lapply(dl_cv_ln, function(x){return(x@coords)})),
                        proj4string = crs(cv))

# extract all data into one big data frame
df_pts <- data.frame(DGBS = do.call(c, lapply(dl_cv_ln, function(x){return(x@data$DGBS)})))

# combine into a space time object (STIDF) for `spacetime`
stidf <- STIDF(sp = sp_pts, time = tt, data = df_pts)

# plot
stplot(stidf, number = 12, main="random spatio-temporal noise")
```


```{r}
# create a regular 20 x 20 x 10 grid of prediction locations:
grd = as(SpatialGrid(GridTopology(c(0.025,0.025), c(.05, .05), c(20,20))), "SpatialPixels")

r <- raster(cv)                 # create a template raster to interpolate over
res(r) <- 5000                  # > township resolution: 6 miles = 9656.06 meters
grd <- as(r, "SpatialGrid")     # convert raster to spatial grid object
grd <- as(grd, "SpatialPixels") # make into spatialpixels object

tgrd = seq(min(tt), max(tt), length.out = 12)

stf = STF(grd, tgrd)

# define a variogram model
sumMetricModel <- vgmST("sumMetric",
                        space=vgm(1/6, "Sph", 0.25, 1/60),
                        time =vgm(2/6, "Exp",  1e5, 1/60),
                        joint=vgm(0.4, "Exp", 0.3, 0.1),
                        stAni=1/1e6)

attr(sumMetricModel, "temporal unit") <- "secs"

dg <- data.frame(spacelag=rep(c(0.001,1:10)/10,6), 
                 timelag=rep(0:5*50e3, each=11))

wireframe(model~spacelag+timelag,
          variogramSurface(sumMetricModel, dist_grid = dg),
          scales=list(arrows=F),
          drape=T, col.regions=bpy.colors(),
          zlim=c(0,1.2),
          main="imposed sum-metric model")
```

```{r}
  kp <- predict(m2, g)                     # OK prediction over grid
  kp@data$var1.pred <- exp(kp@data$var1.pred) # exponentiate the prediciton
  kp@data$var1.var  <- exp(kp@data$var1.var)  # exponentiate the variance
  ok <- brick(kp)                          # spatialgrid df -> raster brick obj.
  ok <- mask(ok, cv)                       # mask to cv extent
  names(ok) <- c('Prediction', 'Variance') # name the raster layers in brick
```


Perform the kirging and exmaine prediciton and variance.
```{r}
locKrig <- krigeST(z~1, stidf, stf, sumMetricModel, nmax=50, computeVar = T)
stplot(locKrig[,,"var1.pred"], col.regions=bpy.colors(), scales=list(draw=T))
stplot(locKrig[,,"var1.var"], col.regions=bpy.colors(), scales=list(draw=T))
```










# following https://www.r-bloggers.com/spatio-temporal-kriging-in-r/
```{r}
zerodist(sp_pts) -> dupl
data.frame(df_pts[-dupl[,2]]) -> ozoneDF
ozoneTM <- tt[-dupl[,2]]
timeDF <- STIDF(sp_pts,ozoneTM,data=ozoneDF)
stplot(timeDF)

var <- variogramST(DGBS~1,data=timeDF,tunit="weeks",assumeRegular=F,na.omit=T) 
plot(var,map=F) 
```
